% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant.R
\name{ants}
\alias{ants}
\title{Ants: Get antonyms for many words}
\usage{
ants(words, n_words = -1)
}
\arguments{
\item{words}{word, character - many words you want to find antonyms for}

\item{n_words}{integer, the number of words to look for. An integer Default is all words}
}
\value{
named list of antonyms
}
\description{
Ants: Get antonyms for many words
}
\examples{
ants(c("good", "evil"), 10)
}
