% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{fetch_and_create}
\alias{fetch_and_create}
\title{Fetch the wrods list and create the R data if it doesn't exist on file}
\usage{
fetch_and_create(words_filename = "~/words.txt",
  syn_filename = "~/syn.rds")
}
\arguments{
\item{words_filename}{location of Moby words.txt file}

\item{syn_filename}{location of processed syn.rds file}
}
\description{
Fetch the wrods list and create the R data if it doesn't exist on file
}
