% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-thesaurus.R
\name{get_thesaurus}
\alias{get_thesaurus}
\title{Install the thesaurus from moby website}
\usage{
get_thesaurus()
}
\value{
nothing - it installs the text file
}
\description{
Installs the thesaurus from the \href{https://github.com/words/moby/blob/master/words.txt}{moby website here}
}
\examples{
\dontrun{
 get_thesaurus()
}
}
