% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom.R
\name{hom}
\alias{hom}
\title{Hom: Get homophones for one word}
\usage{
hom(word, n_words = -1L, algos = "dbm")
}
\arguments{
\item{word}{character, one word you want to find homophones for.}

\item{n_words}{integer, the number of words to look for. An integer Default is all words}

\item{algos}{one or more of c('dbm', 'dbmalt', 'soundex').}
}
\value{
homophones of the word you entered. Returns \code{character(0)} if word not found.
}
\description{
Homophones are determined by comparing the phonetic encoding of words.  If
they have the same phonetic encoding, then they are considered to sound
alike.
}
\details{
The algorithms available for phonetic encoding are:
\itemize{
\item{\code{dbm} - double metaphone (matching primary encodings). \url{https://en.wikipedia.org/wiki/Metaphone#Double_Metaphone}}
\item{\code{dbmalt} - double metaphone (matching primary encodings with secondary encodings)}
\item{\code{soundex} - \url{https://en.wikipedia.org/wiki/Soundex}}
}
}
\examples{
hom("good")
hom("evil", 25)
# No words for spelling mistakes
hom("spolling misteak")
}
