% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom.R
\name{homs}
\alias{homs}
\title{homs: Get homophones for many words}
\usage{
homs(words, n_words = -1, algos = "dbm")
}
\arguments{
\item{words}{word, character - many words you want to find homophones for}

\item{n_words}{integer, the number of words to look for. An integer Default is all words}

\item{algos}{one or more of c('dbm', 'dbmalt', 'soundex').}
}
\value{
named list of homophones. Returns \code{character(0)} if word not found.
}
\description{
homs: Get homophones for many words
}
\examples{
homs(c("good", "evil"), 10)
# No words for spelling mistakes
homs("spolling misteak")
}
