% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-words.R
\name{n_words}
\alias{n_words}
\title{Return the number of words in a string}
\usage{
n_words(x)
}
\arguments{
\item{x}{a string}
}
\value{
vector of numbers
}
\description{
Return the number of words in a string
}
\examples{
syn_end <- syn("end")
syn_end_words <- n_words(syn_end)
syn_end[syn_end_words <= 1]
}
