% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-thesaurus.R
\name{parse_thesaurus}
\alias{parse_thesaurus}
\title{Parse the Moby thesaurus word list into a R list and save to file}
\usage{
parse_thesaurus(words_filename = "~/words.txt",
  syn_filename = "~/syn.rds")
}
\arguments{
\item{words_filename}{the location of 'words.txt' from moby thesaurus}

\item{syn_filename}{the destiniation file}
}
\description{
The word list as downloaded is a raw text file with one index word per line.
Following each index word is a comma-separated list of synonyms.
}
\details{
This code splits that original text file into lines, breaks up the words
from each line (by splitting on commas) and creates a named list of
character vectors.
}
