% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.R
\name{syns}
\alias{syns}
\title{Syns: Get synonyms for many words}
\usage{
syns(words, n_words = -1)
}
\arguments{
\item{words}{word, character - many words you want to find synonyms for}

\item{n_words}{integer, the number of words to look for. An integer Default is all words}
}
\value{
named list of synonyms. Returns \code{character(0)} if word not found.
}
\description{
Syns: Get synonyms for many words
}
\examples{
syns(c("good", "evil"), 10)
# No words for spelling mistakes
syns("spolling misteak")
}
