% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant.R
\name{ant}
\alias{ant}
\title{Ant: Get antonyms for one word}
\usage{
ant(word, n_words = -1)
}
\arguments{
\item{word}{character, one word you want to find antonyms for}

\item{n_words}{integer, the number of words to look for. An integer Default is all words}
}
\value{
antonyms of the word you entered
}
\description{
Ant: Get antonyms for one word
}
\examples{
# put some examples here
ant("good")
ant("evil", 25)
}
