% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-thesaurus.R
\name{parse_thesaurus}
\alias{parse_thesaurus}
\title{Parse the Moby words list}
\usage{
parse_thesaurus(words_filename = "~/words.txt",
  syn_filename = "~/syn.rds")
}
\arguments{
\item{words_filename}{the location of 'words.txt' from moby thesaurus}

\item{syn_filename}{the destiniation file}
}
\description{
Parse the Moby words list
}
